back on searchlight functionality
