+++ app/views/users/show.html.erb
@@ -6,33 +6,18 @@
     <%= render partial: "nav" %>
   </div>
   <div class="">
-
-
-  </div>
-  <table class="table">
-    <tr id="table-header">
-      <th>Idea</th>
-      <th>Category</th>
-      <th>Created</th>
-      <th>Updated</th>
-      <th>Images</th>
-      <th></th>
-      <th></th>
-    </tr>
     <% @ideas.each do |idea| %>
-    <tr>
-      <td><%= idea.body %></td>
-      <td><%= idea.category.classification %></td>
-      <td><%= idea.days_ago %></td>
-      <td><%= idea.updated_date_and_time %></td>
-      <td><% idea.images.each do |image| %>
-        <%= image.name %>
-        <%= image_tag image.image.url unless image.image.url.nil?  %>
-        <% end %></td>
-        <td><%= link_to "Edit", edit_user_idea_path(current_user, idea) %></td>
-        <td><%= link_to "Delete", user_idea_path(current_user, idea), method: :delete, data: { confirm: 'Are you sure?' } %></td>
-      </tr>
-      <% end %>
-    </table>
-    <%= will_paginate @ideas %>
+      <section>
+        <%= idea.body %>
+        <%= idea.category.classification %>
+        <%= idea.time_ago %>
+        <% idea.images.each do |image| %>
+          <%= image_tag image.image.url unless image.image.url.nil?  %>
+        <% end %>
+        <%= link_to "Edit", edit_user_idea_path(current_user, idea) %>
+        <%= link_to "Edit", edit_user_idea_path(current_user, idea) %>
+      </section>
+      <%= will_paginate @ideas %>
+    <% end %>
+  </div>
 </section>
